+++ src/index.js
@@ -44,6 +44,10 @@ export default class {
       'postCreate'
     ]
 
+    this.interceptors = options.interceptors || {
+      postSaveByAction: (() => {})
+    }
+
     // Types
     this.types = options.types || [
       'CREATE',
@@ -138,7 +142,7 @@ export default class {
     const has = type => types.includes(type)
 
     // Params
-    const params = [module, collection, this.postSaveByAction]
+    const params = [module, collection, this.interceptors]
 
     return {
       namespaced: true,