+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 # 2.4.0 2021-08-12
 
 ### Added
-- added new option `postSaveByAction` callback for intercepting all success execution of `create` and `update` action.
+- added new option `interceptors` with `postSaveByAction` callback for intercepting all success execution of `create` and `update` action.
 
 # 2.3.0 2021-07-29